Cambios realizados el 03 de Marzo de 2023

+++ database/migrations/2022_11_03_234433_create_controles_table.php
@@ -17,7 +17,7 @@ class CreateControlesTable extends Migration
             $table->bigIncrements('id');
             $table->string('user_id')->required();
             $table->integer('id_cliente')->unsigned(); 
-            $table->integer('id_historia')->nullable(); 
+            $table->bigInteger('id_historia')->nullable(); 
             $table->string('num_control',3)->nullable();
             $table->string('peso',3)->nullable();
             $table->string('abd',3)->nullable();
